CR feedback and refactoring of ConversionsBase
