+++ src/Compilers/CSharp/Portable/Binder/Semantics/Conversions/ConversionsBase.cs
@@ -25,9 +25,9 @@ namespace Microsoft.CodeAnalysis.CSharp
             _currentRecursionDepth = currentRecursionDepth;
         }
 
-        protected abstract ConversionsBase CreateInstance(int currentRecursionDepth);
-
         public abstract Conversion GetMethodGroupConversion(BoundMethodGroup source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics);
+
+        protected abstract ConversionsBase CreateInstance(int currentRecursionDepth);
 
         protected abstract Conversion GetInterpolatedStringConversion(BoundInterpolatedString source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics);
 
@@ -36,94 +36,130 @@ namespace Microsoft.CodeAnalysis.CSharp
         protected abstract Conversion GetExplicitTupleLiteralConversion(BoundTupleLiteral source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics, bool forCast);
 
         /// <summary>
-        /// Attempt a quick classification of builtin conversions.  As result of "no conversion"
-        /// means that there is no built-in conversion, though there still may be a user-defined
-        /// conversion if compiling against a custom mscorlib.
+        /// Determines if the source expression is convertible to the destination type via
+        /// any built-in or user-defined implicit conversion.
         /// </summary>
-        public static Conversion FastClassifyConversion(TypeSymbol source, TypeSymbol target)
+        public Conversion ClassifyImplicitConversionFromExpression(BoundExpression sourceExpression, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
         {
-            ConversionKind convKind = ConversionEasyOut.ClassifyConversion(source, target);
-            if (convKind != ConversionKind.ImplicitNullable && convKind != ConversionKind.ExplicitNullable)
+            Debug.Assert(sourceExpression != null);
+            Debug.Assert((object)destination != null);
+
+            var sourceType = sourceExpression.Type;
+
+            //PERF: identity conversion is by far the most common implicit conversion, check for that first
+            if ((object)sourceType != null && HasIdentityConversion(sourceType, destination))
             {
-                return Conversion.GetTrivialConversion(convKind);
+                return Conversion.Identity;
             }
 
-            return Conversion.MakeNullableConversion(convKind, FastClassifyConversion(source.StrippedType(), target.StrippedType()));
+            Conversion conversion = ClassifyImplicitBuiltInConversionFromExpression(sourceExpression, sourceType, destination, ref useSiteDiagnostics);
+            if (conversion.Exists)
+            {
+                return conversion;
+            }
+
+            if ((object)sourceType != null)
+            {
+                // Try using the short-circuit "fast-conversion" path.
+                Conversion fastConversion = FastClassifyConversion(sourceType, destination);
+                if (fastConversion.Exists)
+                {
+                    return fastConversion.IsImplicit ? fastConversion : Conversion.NoConversion;
+                }
+                else
+                {
+                    conversion = ClassifyImplicitBuiltInConversionSlow(sourceType, destination, ref useSiteDiagnostics);
+                    if (conversion.Exists)
+                    {
+                        return conversion;
+                    }
+                }
+            }
+
+            return GetImplicitUserDefinedConversion(sourceExpression, sourceType, destination, ref useSiteDiagnostics);
         }
 
         /// <summary>
-        /// IsBaseInterface returns true if baseType is on the base interface list of derivedType or
-        /// any base class of derivedType. It may be on the base interface list either directly or
-        /// indirectly.
-        /// * baseType must be an interface.
-        /// * type parameters do not have base interfaces. (They have an "effective interface list".)
-        /// * an interface is not a base of itself.
-        /// * this does not check for variance conversions; if a type inherits from
-        ///   IEnumerable&lt;string> then IEnumerable&lt;object> is not a base interface.
+        /// Determines if the source type is convertible to the destination type via
+        /// any built-in or user-defined implicit conversion.
         /// </summary>
-        public static bool IsBaseInterface(TypeSymbol baseType, TypeSymbol derivedType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        public Conversion ClassifyImplicitConversionFromType(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
         {
-            Debug.Assert((object)baseType != null);
-            Debug.Assert((object)derivedType != null);
-            if (!baseType.IsInterfaceType())
+            Debug.Assert((object)source != null);
+            Debug.Assert((object)destination != null);
+
+            //PERF: identity conversions are very common, check for that first.
+            if (HasIdentityConversion(source, destination))
             {
-                return false;
+                return Conversion.Identity;
             }
 
-            var d = derivedType as NamedTypeSymbol;
-            if ((object)d == null)
+            // Try using the short-circuit "fast-conversion" path.
+            Conversion fastConversion = FastClassifyConversion(source, destination);
+            if (fastConversion.Exists)
             {
-                return false;
+                return fastConversion.IsImplicit ? fastConversion : Conversion.NoConversion;
             }
-
-            foreach (var iface in d.AllInterfacesWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics))
+            else
             {
-                if (HasIdentityConversion(iface, baseType))
+                Conversion conversion = ClassifyImplicitBuiltInConversionSlow(source, destination, ref useSiteDiagnostics);
+                if (conversion.Exists)
                 {
-                    return true;
+                    return conversion;
                 }
             }
 
-            return false;
+            return GetImplicitUserDefinedConversion(null, source, destination, ref useSiteDiagnostics);
         }
 
-        // IsBaseClassOfClass determines whether the purported derived type is a class, and if so,
-        // if the purported base type is one of its base classes. 
-        public static bool IsBaseClassOfClass(TypeSymbol baseType, TypeSymbol derivedType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        /// <summary>
+        /// Determines if the source expression of given type is convertible to the destination type via
+        /// any built-in or user-defined conversion.
+        /// 
+        /// This helper is used in rare cases involving synthesized expressions where we know the type of an expression, but do not have the actual expression.
+        /// The reason for this helper (as opposed to ClassifyConversionFromType) is that conversions from expressions could be different
+        /// from conversions from type. For example expressions of dynamic type are implicitly convertable to any type, while dynamic type itself is not.
+        /// </summary>
+        public Conversion ClassifyConversionFromExpressionType(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
         {
-            Debug.Assert((object)baseType != null);
-            Debug.Assert((object)derivedType != null);
-            return derivedType.IsClassType() && IsBaseClass(derivedType, baseType, ref useSiteDiagnostics);
+            Debug.Assert((object)source != null);
+            Debug.Assert((object)destination != null);
+
+            // since we are converting from expression, we may have implicit dynamic conversion
+            if (HasImplicitDynamicConversionFromExpression(source, destination))
+            {
+                return Conversion.ImplicitDynamic;
+            }
+
+            return ClassifyConversionFromType(source, destination, ref useSiteDiagnostics);
         }
 
-        // IsBaseClass returns true if and only if baseType is a base class of derivedType, period.
-        //
-        // * interfaces do not have base classes. (Structs, enums and classes other than object do.)
-        // * a class is not a base class of itself
-        // * type parameters do not have base classes. (They have "effective base classes".)
-        // * all base classes must be classes
-        // * dynamics are removed; if we have class D : B<dynamic> then B<object> is a 
-        //   base class of D. However, dynamic is never a base class of anything.
-        public static bool IsBaseClass(TypeSymbol derivedType, TypeSymbol baseType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        /// <summary>
+        /// Determines if the source expression is convertible to the destination type via
+        /// any conversion: implicit, explicit, user-defined or built-in.
+        /// </summary>
+        /// <remarks>
+        /// It is rare but possible for a source expression to be convertible to a destination type
+        /// by both an implicit user-defined conversion and a built-in explicit conversion.
+        /// In that circumstance, this method classifies the conversion as the implicit conversion or explicit depending on "forCast"
+        /// </remarks>
+        public Conversion ClassifyConversionFromExpression(BoundExpression sourceExpression, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics, bool forCast = false)
         {
-            Debug.Assert((object)derivedType != null);
-            Debug.Assert((object)baseType != null);
+            Debug.Assert(sourceExpression != null);
+            Debug.Assert((object)destination != null);
 
-            // A base class has got to be a class. The derived type might be a struct, enum, or delegate.
-            if (!baseType.IsClassType())
+            if (forCast)
             {
-                return false;
+                return ClassifyConversionFromExpressionForCast(sourceExpression, destination, ref useSiteDiagnostics);
             }
 
-            for (TypeSymbol b = derivedType.BaseTypeWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics); (object)b != null; b = b.BaseTypeWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics))
+            var result = ClassifyImplicitConversionFromExpression(sourceExpression, destination, ref useSiteDiagnostics);
+            if (result.Exists)
             {
-                if (HasIdentityConversion(b, baseType))
-                {
-                    return true;
-                }
+                return result;
             }
 
-            return false;
+            return ClassifyExplicitOnlyConversionFromExpression(sourceExpression, destination, ref useSiteDiagnostics, forCast: false);
         }
 
         /// <summary>
@@ -133,12 +169,17 @@ namespace Microsoft.CodeAnalysis.CSharp
         /// <remarks>
         /// It is rare but possible for a source type to be convertible to a destination type
         /// by both an implicit user-defined conversion and a built-in explicit conversion.
-        /// In that circumstance, this method classifies the conversion as the implicit conversion.
+        /// In that circumstance, this method classifies the conversion as the implicit conversion or explicit depending on "forCast"
         /// </remarks>
-        public Conversion ClassifyConversion(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics, bool builtinOnly = false)
+        public Conversion ClassifyConversionFromType(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics, bool forCast = false)
         {
             Debug.Assert((object)source != null);
             Debug.Assert((object)destination != null);
+
+            if (forCast)
+            {
+                return ClassifyConversionFromTypeForCast(source, destination, ref useSiteDiagnostics);
+            }
 
             // Try using the short-circuit "fast-conversion" path.
             Conversion fastConversion = FastClassifyConversion(source, destination);
@@ -155,22 +196,76 @@ namespace Microsoft.CodeAnalysis.CSharp
                 }
             }
 
-            if (!builtinOnly)
-            {
-                Conversion conversion1 = GetImplicitUserDefinedConversion(null, source, destination, ref useSiteDiagnostics);
-                if (conversion1.Exists)
-                {
-                    return conversion1;
-                }
-            }
-
-            var conversion = ClassifyExplicitBuiltInOnlyConversion(source, destination, ref useSiteDiagnostics, forCast: false);
+            Conversion conversion = GetImplicitUserDefinedConversion(null, source, destination, ref useSiteDiagnostics);
             if (conversion.Exists)
             {
                 return conversion;
             }
 
-            return builtinOnly ? conversion : GetExplicitUserDefinedConversion(null, source, destination, ref useSiteDiagnostics);
+            conversion = ClassifyExplicitBuiltInOnlyConversion(source, destination, ref useSiteDiagnostics, forCast: false);
+            if (conversion.Exists)
+            {
+                return conversion;
+            }
+
+            return GetExplicitUserDefinedConversion(null, source, destination, ref useSiteDiagnostics);
+        }
+
+        /// <summary>
+        /// Determines if the source expression is convertible to the destination type via
+        /// any conversion: implicit, explicit, user-defined or built-in.
+        /// </summary>
+        /// <remarks>
+        /// It is rare but possible for a source expression to be convertible to a destination type
+        /// by both an implicit user-defined conversion and a built-in explicit conversion.
+        /// In that circumstance, this method classifies the conversion as the built-in conversion.
+        /// 
+        /// An implicit conversion exists from an expression of a dynamic type to any type.
+        /// An explicit conversion exists from a dynamic type to any type. 
+        /// When casting we prefer the explicit conversion.
+        /// </remarks>
+        private Conversion ClassifyConversionFromExpressionForCast(BoundExpression source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            Debug.Assert(source != null);
+            Debug.Assert((object)destination != null);
+
+            Conversion implicitConversion = ClassifyImplicitConversionFromExpression(source, destination, ref useSiteDiagnostics);
+            if (implicitConversion.Exists && !ExplicitConversionMayDifferFromImplicit(implicitConversion))
+            {
+                return implicitConversion;
+            }
+
+            Conversion explicitConversion = ClassifyExplicitOnlyConversionFromExpression(source, destination, ref useSiteDiagnostics, forCast: true);
+            if (explicitConversion.Exists)
+            {
+                return explicitConversion;
+            }
+
+            // It is possible for a user-defined conversion to be unambiguous when considered as
+            // an implicit conversion and ambiguous when considered as an explicit conversion.
+            // The native compiler does not check to see if a cast could be successfully bound as
+            // an unambiguous user-defined implicit conversion; it goes right to the ambiguous
+            // user-defined explicit conversion and produces an error. This means that in
+            // C# 5 it is possible to have:
+            //
+            // Y y = new Y();
+            // Z z1 = y;
+            // 
+            // succeed but
+            //
+            // Z z2 = (Z)y;
+            //
+            // fail.
+            //
+            // However, there is another interesting wrinkle. It is possible for both
+            // an implicit user-defined conversion and an explicit user-defined conversion
+            // to exist and be unambiguous. For example, if there is an implicit conversion
+            // double-->C and an explicit conversion from int-->C, and the user casts a short
+            // to C, then both the implicit and explicit conversions are applicable and
+            // unambiguous. The native compiler in this case prefers the explicit conversion,
+            // and for backwards compatibility, we match it.
+
+            return implicitConversion.Exists ? implicitConversion : Conversion.NoConversion;
         }
 
         /// <summary>
@@ -182,7 +277,7 @@ namespace Microsoft.CodeAnalysis.CSharp
         /// by both an implicit user-defined conversion and a built-in explicit conversion.
         /// In that circumstance, this method classifies the conversion as the built-in conversion.
         /// </remarks>
-        public Conversion ClassifyConversionForCast(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        private Conversion ClassifyConversionFromTypeForCast(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
         {
             Debug.Assert((object)source != null);
             Debug.Assert((object)destination != null);
@@ -237,6 +332,45 @@ namespace Microsoft.CodeAnalysis.CSharp
         }
 
         /// <summary>
+        /// Attempt a quick classification of builtin conversions.  As result of "no conversion"
+        /// means that there is no built-in conversion, though there still may be a user-defined
+        /// conversion if compiling against a custom mscorlib.
+        /// </summary>
+        public static Conversion FastClassifyConversion(TypeSymbol source, TypeSymbol target)
+        {
+            ConversionKind convKind = ConversionEasyOut.ClassifyConversion(source, target);
+            if (convKind != ConversionKind.ImplicitNullable && convKind != ConversionKind.ExplicitNullable)
+            {
+                return Conversion.GetTrivialConversion(convKind);
+            }
+
+            return Conversion.MakeNullableConversion(convKind, FastClassifyConversion(source.StrippedType(), target.StrippedType()));
+        }
+
+        public Conversion ClassifyBuiltInConversion(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            Debug.Assert((object)source != null);
+            Debug.Assert((object)destination != null);
+
+            // Try using the short-circuit "fast-conversion" path.
+            Conversion fastConversion = FastClassifyConversion(source, destination);
+            if (fastConversion.Exists)
+            {
+                return fastConversion;
+            }
+            else
+            {
+                Conversion conversion = ClassifyImplicitBuiltInConversionSlow(source, destination, ref useSiteDiagnostics);
+                if (conversion.Exists)
+                {
+                    return conversion;
+                }
+            }
+
+            return ClassifyExplicitBuiltInOnlyConversion(source, destination, ref useSiteDiagnostics, forCast: false);
+        }
+
+        /// <summary>
         /// Determines if the source type is convertible to the destination type via
         /// any standard implicit or standard explicit conversion.
         /// </summary>
@@ -282,58 +416,6 @@ namespace Microsoft.CodeAnalysis.CSharp
             }
 
             return Conversion.NoConversion;
-        }
-
-        private Conversion DeriveStandardExplicitFromOppositeStandardImplicitConversion(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
-        {
-            var oppositeConversion = ClassifyStandardImplicitConversion(destination, source, ref useSiteDiagnostics);
-            Conversion impliedExplicitConversion;
-
-            switch (oppositeConversion.Kind)
-            {
-                case ConversionKind.Identity:
-                    impliedExplicitConversion = Conversion.Identity;
-                    break;
-                case ConversionKind.ImplicitNumeric:
-                    impliedExplicitConversion = Conversion.ExplicitNumeric;
-                    break;
-                case ConversionKind.ImplicitReference:
-                    impliedExplicitConversion = Conversion.ExplicitReference;
-                    break;
-                case ConversionKind.Boxing:
-                    impliedExplicitConversion = Conversion.Unboxing;
-                    break;
-                case ConversionKind.NoConversion:
-                    impliedExplicitConversion = Conversion.NoConversion;
-                    break;
-                case ConversionKind.PointerToVoid:
-                    impliedExplicitConversion = Conversion.PointerToPointer;
-                    break;
-
-                case ConversionKind.ImplicitTuple:
-                    // only implicit tuple conversions are standard conversions, 
-                    // having implicit conversion in the other direction does not help here.
-                    impliedExplicitConversion = Conversion.NoConversion;
-                    break;
-
-                case ConversionKind.ImplicitNullable:
-                    var strippedSource = source.StrippedType();
-                    var strippedDestination = destination.StrippedType();
-                    var underlyingConversion = DeriveStandardExplicitFromOppositeStandardImplicitConversion(strippedSource, strippedDestination, ref useSiteDiagnostics);
-
-                    // the opposite underlying conversion may not exist 
-                    // for example if underlying conversion is implicit tuple
-                    impliedExplicitConversion = underlyingConversion.Exists ?
-                        Conversion.MakeNullableConversion(ConversionKind.ExplicitNullable, underlyingConversion) :
-                        Conversion.NoConversion;
-
-                    break;
-
-                default:
-                    throw ExceptionUtilities.UnexpectedValue(oppositeConversion.Kind);
-            }
-
-            return impliedExplicitConversion;
         }
 
         internal Conversion ClassifyStandardImplicitConversion(BoundExpression sourceExpression, TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
@@ -383,7 +465,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             return Conversion.NoConversion;
         }
 
-        internal Conversion ClassifyStandardImplicitConversion(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        private Conversion ClassifyStandardImplicitConversion(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
         {
             Debug.Assert((object)source != null);
             Debug.Assert((object)destination != null);
@@ -446,7 +528,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
             //if (HasImplicitDynamicConversion(source, destination))
             //{
-            //    return Conversion.Dynamic;
+            //    return Conversion.ImplicitDynamic;
             //}
 
             return Conversion.NoConversion;
@@ -456,42 +538,6 @@ namespace Microsoft.CodeAnalysis.CSharp
         {
             var conversionResult = AnalyzeImplicitUserDefinedConversions(sourceExpression, source, destination, ref useSiteDiagnostics);
             return new Conversion(conversionResult, isImplicit: true);
-        }
-
-        /// <summary>
-        /// Determines if the source type is convertible to the destination type via
-        /// any user-defined or built-in implicit conversion.
-        /// </summary>
-        /// <remarks>
-        /// Not all built-in explicit conversions are standard explicit conversions.
-        /// </remarks>
-        public Conversion ClassifyImplicitConversion(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
-        {
-            Debug.Assert((object)source != null);
-            Debug.Assert((object)destination != null);
-
-            //PERF: identity conversions are very common, check for that first.
-            if (HasIdentityConversion(source, destination))
-            {
-                return Conversion.Identity;
-            }
-
-            // Try using the short-circuit "fast-conversion" path.
-            Conversion fastConversion = FastClassifyConversion(source, destination);
-            if (fastConversion.Exists)
-            {
-                return fastConversion.IsImplicit ? fastConversion : Conversion.NoConversion;
-            }
-            else
-            {
-                Conversion conversion = ClassifyImplicitBuiltInConversionSlow(source, destination, ref useSiteDiagnostics);
-                if (conversion.Exists)
-                {
-                    return conversion;
-                }
-            }
-
-            return GetImplicitUserDefinedConversion(null, source, destination, ref useSiteDiagnostics);
         }
 
         private Conversion ClassifyExplicitBuiltInOnlyConversion(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics, bool forCast)
@@ -574,6 +620,684 @@ namespace Microsoft.CodeAnalysis.CSharp
             return new Conversion(conversionResult, isImplicit: false);
         }
 
+        private Conversion DeriveStandardExplicitFromOppositeStandardImplicitConversion(TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            var oppositeConversion = ClassifyStandardImplicitConversion(destination, source, ref useSiteDiagnostics);
+            Conversion impliedExplicitConversion;
+
+            switch (oppositeConversion.Kind)
+            {
+                case ConversionKind.Identity:
+                    impliedExplicitConversion = Conversion.Identity;
+                    break;
+                case ConversionKind.ImplicitNumeric:
+                    impliedExplicitConversion = Conversion.ExplicitNumeric;
+                    break;
+                case ConversionKind.ImplicitReference:
+                    impliedExplicitConversion = Conversion.ExplicitReference;
+                    break;
+                case ConversionKind.Boxing:
+                    impliedExplicitConversion = Conversion.Unboxing;
+                    break;
+                case ConversionKind.NoConversion:
+                    impliedExplicitConversion = Conversion.NoConversion;
+                    break;
+                case ConversionKind.PointerToVoid:
+                    impliedExplicitConversion = Conversion.PointerToPointer;
+                    break;
+
+                case ConversionKind.ImplicitTuple:
+                    // only implicit tuple conversions are standard conversions, 
+                    // having implicit conversion in the other direction does not help here.
+                    impliedExplicitConversion = Conversion.NoConversion;
+                    break;
+
+                case ConversionKind.ImplicitNullable:
+                    var strippedSource = source.StrippedType();
+                    var strippedDestination = destination.StrippedType();
+                    var underlyingConversion = DeriveStandardExplicitFromOppositeStandardImplicitConversion(strippedSource, strippedDestination, ref useSiteDiagnostics);
+
+                    // the opposite underlying conversion may not exist 
+                    // for example if underlying conversion is implicit tuple
+                    impliedExplicitConversion = underlyingConversion.Exists ?
+                        Conversion.MakeNullableConversion(ConversionKind.ExplicitNullable, underlyingConversion) :
+                        Conversion.NoConversion;
+
+                    break;
+
+                default:
+                    throw ExceptionUtilities.UnexpectedValue(oppositeConversion.Kind);
+            }
+
+            return impliedExplicitConversion;
+        }
+
+        /// <summary>
+        /// IsBaseInterface returns true if baseType is on the base interface list of derivedType or
+        /// any base class of derivedType. It may be on the base interface list either directly or
+        /// indirectly.
+        /// * baseType must be an interface.
+        /// * type parameters do not have base interfaces. (They have an "effective interface list".)
+        /// * an interface is not a base of itself.
+        /// * this does not check for variance conversions; if a type inherits from
+        ///   IEnumerable&lt;string> then IEnumerable&lt;object> is not a base interface.
+        /// </summary>
+        public static bool IsBaseInterface(TypeSymbol baseType, TypeSymbol derivedType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            Debug.Assert((object)baseType != null);
+            Debug.Assert((object)derivedType != null);
+            if (!baseType.IsInterfaceType())
+            {
+                return false;
+            }
+
+            var d = derivedType as NamedTypeSymbol;
+            if ((object)d == null)
+            {
+                return false;
+            }
+
+            foreach (var iface in d.AllInterfacesWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics))
+            {
+                if (HasIdentityConversion(iface, baseType))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        // IsBaseClassOfClass determines whether the purported derived type is a class, and if so,
+        // if the purported base type is one of its base classes. 
+        public static bool IsBaseClassOfClass(TypeSymbol baseType, TypeSymbol derivedType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            Debug.Assert((object)baseType != null);
+            Debug.Assert((object)derivedType != null);
+            return derivedType.IsClassType() && IsBaseClass(derivedType, baseType, ref useSiteDiagnostics);
+        }
+
+        // IsBaseClass returns true if and only if baseType is a base class of derivedType, period.
+        //
+        // * interfaces do not have base classes. (Structs, enums and classes other than object do.)
+        // * a class is not a base class of itself
+        // * type parameters do not have base classes. (They have "effective base classes".)
+        // * all base classes must be classes
+        // * dynamics are removed; if we have class D : B<dynamic> then B<object> is a 
+        //   base class of D. However, dynamic is never a base class of anything.
+        public static bool IsBaseClass(TypeSymbol derivedType, TypeSymbol baseType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            Debug.Assert((object)derivedType != null);
+            Debug.Assert((object)baseType != null);
+
+            // A base class has got to be a class. The derived type might be a struct, enum, or delegate.
+            if (!baseType.IsClassType())
+            {
+                return false;
+            }
+
+            for (TypeSymbol b = derivedType.BaseTypeWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics); (object)b != null; b = b.BaseTypeWithDefinitionUseSiteDiagnostics(ref useSiteDiagnostics))
+            {
+                if (HasIdentityConversion(b, baseType))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// returns true when implicit conversion is not necessarily the same as explicit conversion
+        /// </summary>
+        private static bool ExplicitConversionMayDifferFromImplicit(Conversion implicitConversion)
+        {
+            switch (implicitConversion.Kind)
+            {
+                case ConversionKind.ImplicitUserDefined:
+                case ConversionKind.ImplicitDynamic:
+                case ConversionKind.ImplicitTuple:
+                case ConversionKind.ImplicitTupleLiteral:
+                case ConversionKind.ImplicitNullable:
+                    return true;
+
+                default:
+                    return false;
+            }
+        }
+
+        private Conversion ClassifyImplicitBuiltInConversionFromExpression(BoundExpression sourceExpression, TypeSymbol source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            Debug.Assert(sourceExpression != null || (object)source != null);
+            Debug.Assert(sourceExpression == null || (object)sourceExpression.Type == (object)source);
+            Debug.Assert((object)destination != null);
+
+            if (HasImplicitDynamicConversionFromExpression(source, destination))
+            {
+                return Conversion.ImplicitDynamic;
+            }
+
+            // The following conversions only exist for certain form of expressions, 
+            // if we have no expression none if them is applicable.
+            if (sourceExpression == null)
+            {
+                return Conversion.NoConversion;
+            }
+
+            if (HasImplicitEnumerationConversion(sourceExpression, destination))
+            {
+                return Conversion.ImplicitEnumeration;
+            }
+
+            var constantConversion = ClassifyImplicitConstantExpressionConversion(sourceExpression, destination);
+            if (constantConversion.Exists)
+            {
+                return constantConversion;
+            }
+
+            switch (sourceExpression.Kind)
+            {
+                case BoundKind.Literal:
+                    var nullLiteralConversion = ClassifyNullLiteralConversion(sourceExpression, destination);
+                    if (nullLiteralConversion.Exists)
+                    {
+                        return nullLiteralConversion;
+                    }
+                    break;
+
+                case BoundKind.TupleLiteral:
+                    var tupleConversion = ClassifyImplicitTupleLiteralConversion((BoundTupleLiteral)sourceExpression, destination, ref useSiteDiagnostics);
+                    if (tupleConversion.Exists)
+                    {
+                        return tupleConversion;
+                    }
+                    break;
+
+                case BoundKind.UnboundLambda:
+                    if (HasAnonymousFunctionConversion(sourceExpression, destination))
+                    {
+                        return Conversion.AnonymousFunction;
+                    }
+                    break;
+
+                case BoundKind.MethodGroup:
+                    Conversion methodGroupConversion = GetMethodGroupConversion((BoundMethodGroup)sourceExpression, destination, ref useSiteDiagnostics);
+                    if (methodGroupConversion.Exists)
+                    {
+                        return methodGroupConversion;
+                    }
+                    break;
+
+                case BoundKind.InterpolatedString:
+                    Conversion interpolatedStringConversion = GetInterpolatedStringConversion((BoundInterpolatedString)sourceExpression, destination, ref useSiteDiagnostics);
+                    if (interpolatedStringConversion.Exists)
+                    {
+                        return interpolatedStringConversion;
+                    }
+                    break;
+            }
+
+            return Conversion.NoConversion;
+        }
+
+        private static Conversion ClassifyNullLiteralConversion(BoundExpression source, TypeSymbol destination)
+        {
+            Debug.Assert((object)source != null);
+            Debug.Assert((object)destination != null);
+
+            if (!source.IsLiteralNull())
+            {
+                return Conversion.NoConversion;
+            }
+
+            // SPEC: An implicit conversion exists from the null literal to any nullable type. 
+            if (destination.IsNullableType())
+            {
+                // The spec defines a "null literal conversion" specifically as a conversion from
+                // null to nullable type.
+                return Conversion.NullLiteral;
+            }
+
+            // SPEC: An implicit conversion exists from the null literal to any reference type. 
+            // SPEC: An implicit conversion exists from the null literal to type parameter T, 
+            // SPEC: provided T is known to be a reference type. [...] The conversion [is] classified 
+            // SPEC: as implicit reference conversion. 
+
+            if (destination.IsReferenceType)
+            {
+                return Conversion.ImplicitReference;
+            }
+
+            // SPEC: The set of implicit conversions is extended to include...
+            // SPEC: ... from the null literal to any pointer type.
+
+            if (destination is PointerTypeSymbol)
+            {
+                return Conversion.NullToPointer;
+            }
+
+            return Conversion.NoConversion;
+        }
+
+        private static Conversion ClassifyImplicitConstantExpressionConversion(BoundExpression source, TypeSymbol destination)
+        {
+            if (HasImplicitConstantExpressionConversion(source, destination))
+            {
+                return Conversion.ImplicitConstant;
+            }
+
+            // strip nullable from the destination
+            //
+            // the following should work and it is an ImplicitNullable conversion
+            //    int? x = 1;
+            if (destination.Kind == SymbolKind.NamedType)
+            {
+                var nt = (NamedTypeSymbol)destination;
+                if (nt.OriginalDefinition.GetSpecialTypeSafe() == SpecialType.System_Nullable_T &&
+                    HasImplicitConstantExpressionConversion(source, nt.TypeArgumentsNoUseSiteDiagnostics[0]))
+                {
+                    return new Conversion(ConversionKind.ImplicitNullable, Conversion.ImplicitConstantUnderlying);
+                }
+            }
+
+            return Conversion.NoConversion;
+        }
+
+        private Conversion ClassifyImplicitTupleLiteralConversion(BoundTupleLiteral source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            var tupleConversion = GetImplicitTupleLiteralConversion(source, destination, ref useSiteDiagnostics);
+            if (tupleConversion.Exists)
+            {
+                return tupleConversion;
+            }
+
+            // strip nullable from the destination
+            //
+            // the following should work and it is an ImplicitNullable conversion
+            //    (int, double)? x = (1,2);
+            if (destination.Kind == SymbolKind.NamedType)
+            {
+                var nt = (NamedTypeSymbol)destination;
+                if (nt.OriginalDefinition.GetSpecialTypeSafe() == SpecialType.System_Nullable_T)
+                {
+                    var underlyingTupleConversion = GetImplicitTupleLiteralConversion(source, nt.TypeArgumentsNoUseSiteDiagnostics[0], ref useSiteDiagnostics);
+
+                    if (underlyingTupleConversion.Exists)
+                    {
+                        return new Conversion(ConversionKind.ImplicitNullable, ImmutableArray.Create(underlyingTupleConversion));
+                    }
+                }
+            }
+
+            return Conversion.NoConversion;
+        }
+
+        private Conversion ClassifyExplicitTupleLiteralConversion(BoundTupleLiteral source, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics, bool forCast)
+        {
+            var tupleConversion = GetExplicitTupleLiteralConversion(source, destination, ref useSiteDiagnostics, forCast);
+            if (tupleConversion.Exists)
+            {
+                return tupleConversion;
+            }
+
+            // strip nullable from the destination
+            //
+            // the following should work and it is an ExplicitNullable conversion
+            //    var x = ((byte, string)?)(1,null);
+            if (destination.Kind == SymbolKind.NamedType)
+            {
+                var nt = (NamedTypeSymbol)destination;
+                if (nt.OriginalDefinition.GetSpecialTypeSafe() == SpecialType.System_Nullable_T)
+                {
+                    var underlyingTupleConversion = GetExplicitTupleLiteralConversion(source, nt.TypeArgumentsNoUseSiteDiagnostics[0], ref useSiteDiagnostics, forCast);
+
+                    if (underlyingTupleConversion.Exists)
+                    {
+                        return new Conversion(ConversionKind.ExplicitNullable, ImmutableArray.Create(underlyingTupleConversion));
+                    }
+                }
+            }
+
+            return Conversion.NoConversion;
+        }
+
+        internal static bool HasImplicitConstantExpressionConversion(BoundExpression source, TypeSymbol destination)
+        {
+            var constantValue = source.ConstantValue;
+
+            if (constantValue == null)
+            {
+                return false;
+            }
+
+            // An implicit constant expression conversion permits the following conversions:
+
+            // A constant-expression of type int can be converted to type sbyte, byte, short, 
+            // ushort, uint, or ulong, provided the value of the constant-expression is within the
+            // range of the destination type.
+            var specialSource = source.Type.GetSpecialTypeSafe();
+
+            if (specialSource == SpecialType.System_Int32)
+            {
+                //if the constant value could not be computed, be generous and assume the conversion will work
+                int value = constantValue.IsBad ? 0 : constantValue.Int32Value;
+                switch (destination.GetSpecialTypeSafe())
+                {
+                    case SpecialType.System_Byte:
+                        return byte.MinValue <= value && value <= byte.MaxValue;
+                    case SpecialType.System_SByte:
+                        return sbyte.MinValue <= value && value <= sbyte.MaxValue;
+                    case SpecialType.System_Int16:
+                        return short.MinValue <= value && value <= short.MaxValue;
+                    case SpecialType.System_UInt32:
+                        return uint.MinValue <= value;
+                    case SpecialType.System_UInt64:
+                        return (int)ulong.MinValue <= value;
+                    case SpecialType.System_UInt16:
+                        return ushort.MinValue <= value && value <= ushort.MaxValue;
+                    default:
+                        return false;
+                }
+            }
+            else if (specialSource == SpecialType.System_Int64 && destination.GetSpecialTypeSafe() == SpecialType.System_UInt64 && (constantValue.IsBad || 0 <= constantValue.Int64Value))
+            {
+                // A constant-expression of type long can be converted to type ulong, provided the
+                // value of the constant-expression is not negative.
+                return true;
+            }
+
+            return false;
+        }
+
+        private Conversion ClassifyExplicitOnlyConversionFromExpression(BoundExpression sourceExpression, TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics, bool forCast)
+        {
+            Debug.Assert(sourceExpression != null);
+            Debug.Assert((object)destination != null);
+
+            var sourceType = sourceExpression.Type;
+
+            // NB: need to check for explicit tuple literal conversion before checking for explicit conversion from type
+            //     The same literal may have both explicit tuple conversion and explicit tuple literal conversion to the target type.
+            //     They are, however, observably different conversions via the order of argument evaluations and element-wise conversions
+            if (sourceExpression.Kind == BoundKind.TupleLiteral)
+            {
+                Conversion tupleConversion = ClassifyExplicitTupleLiteralConversion((BoundTupleLiteral)sourceExpression, destination, ref useSiteDiagnostics, forCast);
+                if (tupleConversion.Exists)
+                {
+                    return tupleConversion;
+                }
+            }
+
+            if ((object)sourceType != null)
+            {
+                // Try using the short-circuit "fast-conversion" path.
+                Conversion fastConversion = FastClassifyConversion(sourceType, destination);
+                if (fastConversion.Exists)
+                {
+                    return fastConversion;
+                }
+                else
+                {
+                    var conversion = ClassifyExplicitBuiltInOnlyConversion(sourceType, destination, ref useSiteDiagnostics, forCast);
+                    if (conversion.Exists)
+                    {
+                        return conversion;
+                    }
+                }
+            }
+
+            return GetExplicitUserDefinedConversion(sourceExpression, sourceType, destination, ref useSiteDiagnostics);
+        }
+
+        private static bool HasImplicitEnumerationConversion(BoundExpression source, TypeSymbol destination)
+        {
+            Debug.Assert((object)source != null);
+            Debug.Assert((object)destination != null);
+
+            // SPEC: An implicit enumeration conversion permits the decimal-integer-literal 0 to be converted to any enum-type 
+            // SPEC: and to any nullable-type whose underlying type is an enum-type. 
+            //
+            // For historical reasons we actually allow a conversion from any *numeric constant
+            // zero* to be converted to any enum type, not just the literal integer zero.
+
+            bool validType = destination.IsEnumType() ||
+                destination.IsNullableType() && destination.GetNullableUnderlyingType().IsEnumType();
+
+            if (!validType)
+            {
+                return false;
+            }
+
+            var sourceConstantValue = source.ConstantValue;
+            return sourceConstantValue != null &&
+                IsNumericType(source.Type.GetSpecialTypeSafe()) &&
+                IsConstantNumericZero(sourceConstantValue);
+        }
+
+        private static LambdaConversionResult IsAnonymousFunctionCompatibleWithDelegate(UnboundLambda anonymousFunction, TypeSymbol type)
+        {
+            Debug.Assert((object)anonymousFunction != null);
+            Debug.Assert((object)type != null);
+
+            // SPEC: An anonymous-method-expression or lambda-expression is classified as an anonymous function. 
+            // SPEC: The expression does not have a type but can be implicitly converted to a compatible delegate 
+            // SPEC: type or expression tree type. Specifically, a delegate type D is compatible with an 
+            // SPEC: anonymous function F provided:
+
+            var delegateType = (NamedTypeSymbol)type;
+            var invokeMethod = delegateType.DelegateInvokeMethod;
+
+            if ((object)invokeMethod == null || invokeMethod.HasUseSiteError)
+            {
+                return LambdaConversionResult.BadTargetType;
+            }
+
+            var delegateParameters = invokeMethod.Parameters;
+
+            // SPEC: If F contains an anonymous-function-signature, then D and F have the same number of parameters.
+            // SPEC: If F does not contain an anonymous-function-signature, then D may have zero or more parameters 
+            // SPEC: of any type, as long as no parameter of D has the out parameter modifier.
+
+            if (anonymousFunction.HasSignature)
+            {
+                if (anonymousFunction.ParameterCount != invokeMethod.ParameterCount)
+                {
+                    return LambdaConversionResult.BadParameterCount;
+                }
+
+                // SPEC: If F has an explicitly typed parameter list, each parameter in D has the same type 
+                // SPEC: and modifiers as the corresponding parameter in F.
+                // SPEC: If F has an implicitly typed parameter list, D has no ref or out parameters.
+
+                if (anonymousFunction.HasExplicitlyTypedParameterList)
+                {
+                    for (int p = 0; p < delegateParameters.Length; ++p)
+                    {
+                        if (delegateParameters[p].RefKind != anonymousFunction.RefKind(p) ||
+                            !delegateParameters[p].Type.Equals(anonymousFunction.ParameterType(p), ignoreCustomModifiersAndArraySizesAndLowerBounds: true, ignoreDynamic: true))
+                        {
+                            return LambdaConversionResult.MismatchedParameterType;
+                        }
+                    }
+                }
+                else
+                {
+                    for (int p = 0; p < delegateParameters.Length; ++p)
+                    {
+                        if (delegateParameters[p].RefKind != RefKind.None)
+                        {
+                            return LambdaConversionResult.RefInImplicitlyTypedLambda;
+                        }
+                    }
+
+                    // In C# it is not possible to make a delegate type
+                    // such that one of its parameter types is a static type. But static types are 
+                    // in metadata just sealed abstract types; there is nothing stopping someone in
+                    // another language from creating a delegate with a static type for a parameter,
+                    // though the only argument you could pass for that parameter is null.
+                    // 
+                    // In the native compiler we forbid conversion of an anonymous function that has
+                    // an implicitly-typed parameter list to a delegate type that has a static type
+                    // for a formal parameter type. However, we do *not* forbid it for an explicitly-
+                    // typed lambda (because we already require that the explicitly typed parameter not
+                    // be static) and we do not forbid it for an anonymous method with the entire
+                    // parameter list missing (because the body cannot possibly have a parameter that
+                    // is of static type, even though this means that we will be generating a hidden
+                    // method with a parameter of static type.)
+                    //
+                    // We also allow more exotic situations to work in the native compiler. For example,
+                    // though it is not possible to convert x=>{} to Action<GC>, it is possible to convert
+                    // it to Action<List<GC>> should there be a language that allows you to construct 
+                    // a variable of that type.
+                    //
+                    // We might consider beefing up this rule to disallow a conversion of *any* anonymous
+                    // function to *any* delegate that has a static type *anywhere* in the parameter list.
+
+                    for (int p = 0; p < delegateParameters.Length; ++p)
+                    {
+                        if (delegateParameters[p].Type.IsStatic)
+                        {
+                            return LambdaConversionResult.StaticTypeInImplicitlyTypedLambda;
+                        }
+                    }
+                }
+            }
+            else
+            {
+                for (int p = 0; p < delegateParameters.Length; ++p)
+                {
+                    if (delegateParameters[p].RefKind == RefKind.Out)
+                    {
+                        return LambdaConversionResult.MissingSignatureWithOutParameter;
+                    }
+                }
+            }
+
+            // Ensure the body can be converted to that delegate type
+            var bound = anonymousFunction.Bind(delegateType);
+            if (ErrorFacts.PreventsSuccessfulDelegateConversion(bound.Diagnostics))
+            {
+                return LambdaConversionResult.BindingFailed;
+            }
+
+            return LambdaConversionResult.Success;
+        }
+
+        private static LambdaConversionResult IsAnonymousFunctionCompatibleWithExpressionTree(UnboundLambda anonymousFunction, NamedTypeSymbol type)
+        {
+            Debug.Assert((object)anonymousFunction != null);
+            Debug.Assert((object)type != null);
+            Debug.Assert(type.IsExpressionTree());
+
+            // SPEC OMISSION:
+            // 
+            // The C# 3 spec said that anonymous methods and statement lambdas are *convertible* to expression tree
+            // types if the anonymous method/statement lambda is convertible to its delegate type; however, actually
+            // *using* such a conversion is an error. However, that is not what we implemented. In C# 3 we implemented
+            // that an anonymous method is *not convertible* to an expression tree type, period. (Statement lambdas
+            // used the rule described in the spec.)  
+            //
+            // This appears to be a spec omission; the intention is to make old-style anonymous methods not 
+            // convertible to expression trees.
+
+            var delegateType = type.TypeArgumentsNoUseSiteDiagnostics[0];
+            if (!delegateType.IsDelegateType())
+            {
+                return LambdaConversionResult.ExpressionTreeMustHaveDelegateTypeArgument;
+            }
+
+            if (anonymousFunction.Syntax.Kind() == SyntaxKind.AnonymousMethodExpression)
+            {
+                return LambdaConversionResult.ExpressionTreeFromAnonymousMethod;
+            }
+
+            return IsAnonymousFunctionCompatibleWithDelegate(anonymousFunction, delegateType);
+        }
+
+        public static LambdaConversionResult IsAnonymousFunctionCompatibleWithType(UnboundLambda anonymousFunction, TypeSymbol type)
+        {
+            Debug.Assert((object)anonymousFunction != null);
+            Debug.Assert((object)type != null);
+
+            if (type.IsDelegateType())
+            {
+                return IsAnonymousFunctionCompatibleWithDelegate(anonymousFunction, type);
+            }
+            else if (type.IsExpressionTree())
+            {
+                return IsAnonymousFunctionCompatibleWithExpressionTree(anonymousFunction, (NamedTypeSymbol)type);
+            }
+
+            return LambdaConversionResult.BadTargetType;
+        }
+
+        private static bool HasAnonymousFunctionConversion(BoundExpression source, TypeSymbol destination)
+        {
+            Debug.Assert(source != null);
+            Debug.Assert((object)destination != null);
+
+            if (source.Kind != BoundKind.UnboundLambda)
+            {
+                return false;
+            }
+
+            return IsAnonymousFunctionCompatibleWithType((UnboundLambda)source, destination) == LambdaConversionResult.Success;
+        }
+
+        internal Conversion ClassifyImplicitUserDefinedConversionForSwitchGoverningType(TypeSymbol sourceType, out TypeSymbol switchGoverningType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            // SPEC:    The governing type of a switch statement is established by the switch expression.
+            // SPEC:    1) If the type of the switch expression is sbyte, byte, short, ushort, int, uint,
+            // SPEC:       long, ulong, bool, char, string, or an enum-type, or if it is the nullable type
+            // SPEC:       corresponding to one of these types, then that is the governing type of the switch statement. 
+            // SPEC:    2) Otherwise, exactly one user-defined implicit conversion (§6.4) must exist from the
+            // SPEC:       type of the switch expression to one of the following possible governing types:
+            // SPEC:       sbyte, byte, short, ushort, int, uint, long, ulong, char, string, or, a nullable type
+            // SPEC:       corresponding to one of those types
+
+            // NOTE:    We should be called only if (1) is false for source type.
+            Debug.Assert((object)sourceType != null);
+            Debug.Assert(!sourceType.IsValidSwitchGoverningType());
+
+            UserDefinedConversionResult result = AnalyzeImplicitUserDefinedConversionForSwitchGoverningType(sourceType, ref useSiteDiagnostics);
+
+            if (result.Kind == UserDefinedConversionResultKind.Valid)
+            {
+                UserDefinedConversionAnalysis analysis = result.Results[result.Best];
+
+                switchGoverningType = analysis.ToType;
+                Debug.Assert(switchGoverningType.IsValidSwitchGoverningType(isTargetTypeOfUserDefinedOp: true));
+            }
+            else
+            {
+                switchGoverningType = null;
+            }
+
+            return new Conversion(result, isImplicit: true);
+        }
+
+        internal Conversion GetCallerLineNumberConversion(TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            var greenNode = new Syntax.InternalSyntax.LiteralExpressionSyntax(SyntaxKind.NumericLiteralExpression, new Syntax.InternalSyntax.SyntaxToken(SyntaxKind.NumericLiteralToken));
+            var syntaxNode = new LiteralExpressionSyntax(greenNode, null, 0);
+
+            TypeSymbol expectedAttributeType = corLibrary.GetSpecialType(SpecialType.System_Int32);
+            BoundLiteral intMaxValueLiteral = new BoundLiteral(syntaxNode, ConstantValue.Create(int.MaxValue), expectedAttributeType);
+            return ClassifyStandardImplicitConversion(intMaxValueLiteral, expectedAttributeType, destination, ref useSiteDiagnostics);
+        }
+
+        internal bool HasCallerLineNumberConversion(TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            return GetCallerLineNumberConversion(destination, ref useSiteDiagnostics).Exists;
+        }
+
+        internal bool HasCallerInfoStringConversion(TypeSymbol destination, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
+        {
+            TypeSymbol expectedAttributeType = corLibrary.GetSpecialType(SpecialType.System_String);
+            Conversion conversion = ClassifyStandardImplicitConversion(expectedAttributeType, destination, ref useSiteDiagnostics);
+            return conversion.Exists;
+        }
+
         public static bool HasIdentityConversion(TypeSymbol type1, TypeSymbol type2)
         {
             // Spec (6.1.1):
@@ -608,7 +1332,7 @@ namespace Microsoft.CodeAnalysis.CSharp
         public Conversion ConvertExtensionMethodThisArg(TypeSymbol parameterType, TypeSymbol thisType, ref HashSet<DiagnosticInfo> useSiteDiagnostics)
         {
             Debug.Assert((object)thisType != null);
-            var conversion = this.ClassifyImplicitConversion(thisType, parameterType, ref useSiteDiagnostics);
+            var conversion = this.ClassifyImplicitConversionFromType(thisType, parameterType, ref useSiteDiagnostics);
             return IsValidExtensionMethodThisArgConversion(conversion) ? conversion : Conversion.NoConversion;
         }
 
@@ -986,7 +1710,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             var nestedConversions = ArrayBuilder<Conversion>.GetInstance(sourceTypes.Length);
             for (int i = 0; i < sourceTypes.Length; i++)
             {
-                var conversion = ClassifyImplicitConversion(sourceTypes[i], destTypes[i], ref useSiteDiagnostics);
+                var conversion = ClassifyImplicitConversionFromType(sourceTypes[i], destTypes[i], ref useSiteDiagnostics);
                 if (!conversion.Exists)
                 {
                     nestedConversions.Free();
@@ -1014,10 +1738,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             var nestedConversions = ArrayBuilder<Conversion>.GetInstance(sourceTypes.Length);
             for (int i = 0; i < sourceTypes.Length; i++)
             {
-                var conversion = forCast ?
-                    ClassifyConversionForCast(sourceTypes[i], destTypes[i], ref useSiteDiagnostics) :
-                    ClassifyConversionFromType(sourceTypes[i], destTypes[i], ref useSiteDiagnostics);
-
+                var conversion = ClassifyConversionFromType(sourceTypes[i], destTypes[i], ref useSiteDiagnostics, forCast);
                 if (!conversion.Exists)
                 {
                     nestedConversions.Free();
